avoid canonicalization of layout helper read and fixed bug in flat and null-restricted array check snippets

+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/hotspot/GraalHotSpotVMConfig.java
@@ -251,10 +251,10 @@ public class GraalHotSpotVMConfig extends GraalHotSpotVMConfigAccess {
     public final int klassProtoTypeHeaderOffset = getFieldOffset("Klass::_prototype_header", Integer.class, "markWord");
     public final int klassKind = getFieldOffset("Klass::_kind", Integer.class, "Klass::KlassKind const");
     public final int klassFlatArrayKlassKind = getConstant("Klass::FlatArrayKlassKind", Integer.class);
-    public final int flatArrayPattern = getConstant("markWord::flat_array_pattern", Integer.class);
-    public final int flatArrayMaskInPlace = getConstant("markWord::flat_array_mask_in_place", Integer.class);
-    public final int nullFreeArrayPattern = getConstant("markWord::null_free_array_pattern", Integer.class);
-    public final int nullFreeArrayMaskInPlace = getConstant("markWord::null_free_array_mask_in_place", Integer.class);
+    public final long flatArrayPattern = getConstant("markWord::flat_array_pattern", Long.class);
+    public final long flatArrayMaskInPlace = getConstant("markWord::flat_array_mask_in_place", Long.class);
+    public final long nullFreeArrayPattern = getConstant("markWord::null_free_array_pattern", Long.class);
+    public final long nullFreeArrayMaskInPlace = getConstant("markWord::null_free_array_mask_in_place", Long.class);
 
     public final int klassLayoutHelperNeutralValue = getConstant("Klass::_lh_neutral_value", Integer.class);
     public final int layoutHelperLog2ElementSizeShift = getConstant("Klass::_lh_log2_element_size_shift", Integer.class);
